refactor(deprecation,errors): clean up errors.py import hierarchy & move DelimiterError in there
