+++ falcon/cyutil/reader.pyx
@@ -19,7 +19,8 @@ from libc.stdint cimport uint32_t
 import functools
 import io
 
-from falcon.util.reader import DelimiterError, DEFAULT_CHUNK_SIZE, _MAX_JOIN_CHUNKS
+from falcon.errors import DelimiterError
+from falcon.util.reader import DEFAULT_CHUNK_SIZE, _MAX_JOIN_CHUNKS
 
 
 cdef class BufferedReader:
